# docs/swagger.yaml
definitions:
...
  internal.Component:
    properties:
      chart:
        $ref: '#/definitions/internal.ComponentChart'
        type: object
      dependencies:
        items:
          $ref: '#/definitions/internal.Component'
        type: array
      image:
        $ref: '#/definitions/internal.ComponentImage'
        type: object
      name:
        type: string
      parent:
        type: string
      source:
        type: string
      values:
        type: object
    type: object
  internal.ComponentChart:
    properties:
      name:
        type: string
      repository:
        type: string
      version:
        type: string
    type: object
  internal.ComponentImage:
    properties:
      pattern:
        type: string
      repository:
        type: string
      tag:
        type: string
    type: object
...
  v1beta1.ActivePromotion:
    properties:
      spec:
        $ref: '#/definitions/v1beta1.ActivePromotionSpec'
        type: object
      status:
        $ref: '#/definitions/v1beta1.ActivePromotionStatus'
        type: object
    type: object
  v1beta1.ActivePromotionCondition:
    properties:
      lastTransitionTime:
        description: +optional
        type: string
      message:
        description: +optional
        type: string
      reason:
        description: +optional
        type: string
      status:
        type: string
      type:
        type: string
    type: object
...
  v1beta1.ActivePromotionHistorySpec:
    properties:
      activePromotion:
        $ref: '#/definitions/v1beta1.ActivePromotion'
        type: object
      createdAt:
        description: 'TODO: store values file of all components'
        type: string
      isSuccess:
        type: boolean
      teamName:
        type: string
    type: object
  v1beta1.ActivePromotionHistoryStatus:
    type: object
  v1beta1.ActivePromotionSpec:
    properties:
      tearDownDuration:
        description: |-
          TearDownDuration represents duration before tear down the previous active namespace
          +optional
        type: string
    type: object
  v1beta1.ActivePromotionStatus:
    properties:
      activePromotionHistoryName:
        description: |-
          ActivePromotionHistoryName represents created ActivePromotionHistoryName name
          +optional
        type: string
      conditions:
        description: |-
          Conditions contains observations of the resource's state e.g.,
          Queue deployed, being tested
          +optional
          +patchMergeKey=type
          +patchStrategy=merge
        items:
          $ref: '#/definitions/v1beta1.ActivePromotionCondition'
        type: array
      demotionStatus:
        description: |-
          DemotionStatus represents a status of the active demotion
          +optional
        type: string
      destroyTime:
        description: |-
          DestroyTime represents time at which the previous active namespace will be destroyed
          +optional
        type: string
      hasOutdatedComponent:
        description: |-
          HasOutdatedComponent defines whether current active promotion has outdated component or not
          +optional
        type: boolean
      isTimeout:
        description: |-
          IsTimeout defines whether the active promotion has been timeout or not
          +optional
        type: boolean
      message:
        description: |-
          Message defines details about why the active promotion is in this condition
          +optional
        type: string
      outdatedComponents:
        description: |-
          OutdatedComponents represents list of outdated components
          +optional
        items:
          $ref: '#/definitions/v1beta1.OutdatedComponent'
        type: array
      preActiveQueue:
        $ref: '#/definitions/v1beta1.QueueStatus'
        description: |-
          PreActiveQueue represents a pre-active queue status
          +optional
        type: object
      previousActiveNamespace:
        description: |-
          PreviousActiveNamespace represents an active namespace before promoting
          +optional
        type: string
      result:
        description: |-
          Result represents a result of the active promotion
          +optional
        type: string
      rollbackStatus:
        description: |-
          RollbackStatus represents a status of the rollback process
          +optional
        type: string
      startedAt:
        description: |-
          StartedAt represents time at which the active promotion started
          +optional
        type: string
      state:
        description: |-
          ActivePromotionState represents a current state of the active promotion
          +optional
        type: string
      targetNamespace:
        description: |-
          TargetNamespace represents a pre-active namespace
          +optional
        type: string
      updatedAt:
        description: |-
          UpdatedAt represents time at which the active promotion finished
          +optional
        type: string
    type: object
  v1beta1.Credential:
    properties:
      git:
        $ref: '#/definitions/v1beta1.UsernamePasswordCredential'
        description: Git
        type: object
      secretName:
        description: SecretName
        type: string
      teamcity:
        $ref: '#/definitions/v1beta1.UsernamePasswordCredential'
        description: |-
          Teamcity
          +optional
        type: object
    type: object
  v1beta1.GitStorage:
    properties:
      cloneDepth:
        description: |-
          CloneDepth limits fetching to the specified number of commits
          +optional
        type: integer
      cloneTimeout:
        description: |-
          CloneTimeout represents timeout duration of git clone
          +optional
        type: string
      path:
        description: |-
          Path represents a directory path in git repo
          +optional
        type: string
      pullTimeout:
        description: |-
          PullTimeout represents timeout duration of git pull
          +optional
        type: string
      pushTimeout:
        description: |-
          PushTimeout represents timeout duration of git push
          +optional
        type: string
      ref:
        description: |-
          Ref represents git ref
          +optional
        type: string
      url:
        description: URL represents git url
        type: string
    type: object
  v1beta1.Image:
    properties:
      repository:
        type: string
      tag:
        type: string
    type: object
  v1beta1.OutdatedComponent:
    properties:
      currentImage:
        $ref: '#/definitions/v1beta1.Image'
        type: object
      latestImage:
        $ref: '#/definitions/v1beta1.Image'
        type: object
      name:
        type: string
      outdatedDuration:
        type: string
    type: object
  v1beta1.Queue:
    properties:
      spec:
        $ref: '#/definitions/v1beta1.QueueSpec'
        type: object
      status:
        $ref: '#/definitions/v1beta1.QueueStatus'
        type: object
    type: object
  v1beta1.QueueCondition:
    properties:
      lastTransitionTime:
        description: +optional
        type: string
      message:
        description: +optional
        type: string
      reason:
        description: +optional
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  v1beta1.QueueHistory:
    properties:
      spec:
        $ref: '#/definitions/v1beta1.QueueHistorySpec'
        type: object
      status:
        $ref: '#/definitions/v1beta1.QueueHistoryStatus'
        type: object
    type: object
  v1beta1.QueueHistorySpec:
    properties:
      appliedValues:
        $ref: '#/definitions/v1beta1.Values'
        type: object
      createdAt:
        type: string
      isDeploySuccess:
        type: boolean
      isReverify:
        type: boolean
      isTestSuccess:
        type: boolean
      queue:
        $ref: '#/definitions/v1beta1.Queue'
        type: object
      stableComponents:
        items:
          $ref: '#/definitions/v1beta1.StableComponent'
        type: array
    type: object
  v1beta1.QueueHistoryStatus:
    type: object
  v1beta1.QueueSpec:
    properties:
      name:
        description: Name represents Component name
        type: string
      nextProcessAt:
        description: NextProcessAt represents time to wait for process this queue
        type: string
      noOfOrder:
        description: |-
          NoOfOrder defines the position in queue
          lower is will be picked first
        type: integer
      noOfRetry:
        description: |-
          NoOfRetry defines how many times this component has been tested
          +optional
        type: integer
      repository:
        description: Repository represents Docker image repository
        type: string
      teamName:
        description: TeamName represents team owner of the queue
        type: string
      type:
        description: Type represents how we will process this queue
        type: string
      version:
        description: Version represents Docker image tag version
        type: string
    type: object
  v1beta1.QueueStatus:
    properties:
      conditions:
        description: |-
          Conditions contains observations of the resource's state e.g.,
          Queue deployed, being tested
          +optional
          +patchMergeKey=type
          +patchStrategy=merge
        items:
          $ref: '#/definitions/v1beta1.QueueCondition'
        type: array
      createdAt:
        description: CreatedAt represents time when the component has been added to
          queue
        type: string
      imageMissingList:
        description: ImageMissingList defines image missing list
        items:
          $ref: '#/definitions/v1beta1.Image'
        type: array
      kubeZipLog:
        description: KubeZipLog defines log of k8s resources during deployment in
          base64 zip format
        type: string
      nextProcessAt:
        description: NextProcessAt represents time to wait for process this queue
        type: string
      noOfProcessed:
        description: NoOfProcessed represents how many time that this queue had been
          processed
        type: integer
      queueHistoryName:
        description: QueueHistoryName defines name of history of this queue
        type: string
      releaseName:
        description: ReleaseName defines name of helmrelease
        type: string
      startDeployTime:
        description: StartDeployTime represents the time when this queue start deploying
        type: string
      startTestingTime:
        description: StartTestingTime represents the time when this queue start testing
        type: string
      state:
        description: State represents current status of this queue
        type: string
      testRunners:
        $ref: '#/definitions/v1beta1.TestRunner'
        description: TestRunner defines the test runner
        type: object
      updatedAt:
        description: UpdatedAt represents time when the component was processed
        type: string
    type: object
  v1beta1.StableComponent:
    properties:
      spec:
        $ref: '#/definitions/v1beta1.StableComponentSpec'
        type: object
      status:
        $ref: '#/definitions/v1beta1.StableComponentStatus'
        type: object
    type: object
  v1beta1.StableComponentSpec:
    properties:
      name:
        description: Name represents Component name
        type: string
      repository:
        description: Repository represents Docker image repository
        type: string
      version:
        description: Version represents Docker image tag version
        type: string
    type: object
  v1beta1.StableComponentStatus:
    properties:
      createdAt:
        type: string
      updatedAt:
        type: string
    type: object
  v1beta1.StagingCtrl:
    properties:
      endpoint:
        description: Endpoint represents the staging endpoint endpoint.
        type: string
      image:
        description: Image represents image for run staging controller.
        type: string
      isDeploy:
        description: IsDeploy represents flag to deploy staging controller or not.
        type: boolean
      resources:
        type: string
    type: object
  v1beta1.TeamCondition:
    properties:
      lastTransitionTime:
        description: +optional
        type: string
      message:
        description: +optional
        type: string
      reason:
        description: +optional
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  v1beta1.TeamNamespace:
    properties:
      active:
        description: +optional
        type: string
      preActive:
        description: +optional
        type: string
      previousActive:
        description: +optional
        type: string
      staging:
        description: +optional
        type: string
    type: object
  v1beta1.TeamSpec:
    properties:
      credential:
        $ref: '#/definitions/v1beta1.Credential'
        description: |-
          Credential
          +optional
        type: object
      desc:
        description: |-
          Description represents description for this team
          +optional
        type: string
      gitStorage:
        $ref: '#/definitions/v1beta1.GitStorage'
        description: GitStorage represents git as storage for retrieving configuration,
          etc.
        type: object
      owners:
        description: |-
          Owners represents contact point of this team
          +optional
        items:
          type: string
        type: array
      resources:
        description: |-
          Resources represents how many resources per namespace for the team
          +optional
        type: string
      stagingCtrl:
        $ref: '#/definitions/v1beta1.StagingCtrl'
        description: |-
          StagingCtrl represents configuration about the staging controller.
          For easier for developing, debugging and testing purposes
          +optional
        type: object
    type: object
  v1beta1.TeamStatus:
    properties:
      conditions:
        description: |-
          Conditions contains observations of the resource's state e.g.,
          Team namespace is created, destroyed
          +optional
          +patchMergeKey=type
          +patchStrategy=merge
        items:
          $ref: '#/definitions/v1beta1.TeamCondition'
        type: array
      desiredComponentImageCreatedTime:
        description: |-
          DesiredComponentImageCreatedTime represents mapping of desired component image and created time
          map[componentName][repository:tag] = image and createdTime
          +optional
        type: object
      namespace:
        $ref: '#/definitions/v1beta1.TeamNamespace'
        description: +optional
        type: object
      stableComponents:
        description: |-
          StableComponentList represents a list of stable components
          +optional
        items:
          $ref: '#/definitions/v1beta1.StableComponent'
        type: array
    type: object
  v1beta1.Teamcity:
    properties:
      buildID:
        type: string
      buildTypeID:
        type: string
      buildURL:
        type: string
    type: object
  v1beta1.TestRunner:
    properties:
      teamcity:
        $ref: '#/definitions/v1beta1.Teamcity'
        type: object
    type: object
  v1beta1.UsernamePasswordCredential:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  v1beta1.Values:
    additionalProperties:
      type: object
    type: object
  webhook.activePromotion:
    properties:
      inQueues:
        description: +Optional
        example:
        - team1
        - team2
        items:
          type: string
        type: array
      running:
        description: +Optional
        items:
          $ref: '#/definitions/v1beta1.ActivePromotion'
        type: array
    type: object
  webhook.activePromotionHistories:
    items:
      properties:
        spec:
          $ref: '#/definitions/v1beta1.ActivePromotionHistorySpec'
          type: object
        status:
          $ref: '#/definitions/v1beta1.ActivePromotionHistoryStatus'
          type: object
      type: object
    type: array
  webhook.errResp:
    properties:
      error:
        type: string
    type: object
  webhook.githubPushEventJSON:
    properties:
      ref:
        type: string
      repository:
        properties:
          full_name:
            type: string
          name:
            type: string
        type: object
    type: object
...
  webhook.newComponentEventJSON:
    properties:
      component:
        type: string
      repository:
        type: string
    type: object
  webhook.teamActivePromotion:
    properties:
      current:
        $ref: '#/definitions/v1beta1.ActivePromotion'
        description: +Optional
        type: object
      historyNames:
        description: +Optional
        example:
        - team1-20191010-080000
        - team1-20191009-080000
        items:
          type: string
        type: array
    type: object
  webhook.teamComponentsJSON:
    additionalProperties:
      properties:
        chart:
          $ref: '#/definitions/internal.ComponentChart'
          type: object
        dependencies:
          items:
            $ref: '#/definitions/internal.Component'
          type: array
        image:
          $ref: '#/definitions/internal.ComponentImage'
          type: object
        name:
          type: string
        parent:
          type: string
        source:
          type: string
        values:
          type: object
      type: object
    type: object
  webhook.teamEnvConnections:
    properties:
      active:
        description: +optional
        type: object
      preActive:
        description: +optional
        type: object
      staging:
        description: +optional
        type: object
    type: object
  webhook.teamJSON:
    properties:
      active:
        description: +optional
        type: string
      connections:
        $ref: '#/definitions/webhook.teamEnvConnections'
        type: object
      preActive:
        description: +optional
        type: string
      previousActive:
        description: +optional
        type: string
      spec:
        $ref: '#/definitions/v1beta1.TeamSpec'
        type: object
      staging:
        description: +optional
        type: string
      status:
        $ref: '#/definitions/v1beta1.TeamStatus'
        type: object
      teamName:
        type: string
    type: object
  webhook.teamQueueJSON:
    properties:
      current:
        $ref: '#/definitions/v1beta1.Queue'
        description: +Optional
        type: object
      historyNames:
        items:
          type: string
        type: array
      noOfQueue:
        description: +optional
        type: integer
      queues:
        description: +Optional
        items:
          $ref: '#/definitions/v1beta1.Queue'
        type: array
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /activepromotions:
    get:
      description: get current active promotions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webhook.activePromotion'
        "500":
          description: cannot get activepromotions
          schema:
            $ref: '#/definitions/webhook.errResp'
      summary: get current active promotions
      tags:
      - GET
...
  /teams/{team}:
    get:
      description: Returns team information. (namespaces, connections)
      parameters:
      - description: Team name
        in: path
        name: team
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webhook.teamJSON'
        "404":
          description: Team not found
          schema:
            $ref: '#/definitions/webhook.errResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/webhook.errResp'
      summary: Get Team
      tags:
      - GET
  /teams/{team}/activepromotions:
    get:
      description: get active promotions by team name
      parameters:
      - description: Team name
        in: path
        name: team
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webhook.teamActivePromotion'
        "404":
          description: team {team} not found
          schema:
            type: string
        "500":
          description: cannot get activepromotion/activepromotion histories of team
            {team}
          schema:
            type: string
      summary: get active promotions by team name
      tags:
      - GET
  /teams/{team}/activepromotions/histories:
    get:
      description: get active promotion histories by team name
      parameters:
      - description: Team name
        in: path
        name: team
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webhook.activePromotionHistories'
        "400":
          description: team should not be empty
          schema:
            type: string
        "500":
          description: cannot get activepromotion histories of team {team}
          schema:
            type: string
      summary: get active promotion histories by team name
      tags:
      - GET
  /teams/{team}/activepromotions/histories/{history}:
    get:
      description: get active promotion history by team and history name
      parameters:
      - description: Team name
        in: path
        name: team
        required: true
        type: string
      - description: Active promotion history name
        in: path
        name: history
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: team/history should not be empty
          schema:
            type: string
        "404":
          description: activepromotion history {history} of team {team} not found
          schema:
            type: string
        "500":
          description: cannot get activepromotion history {history} of team {team}
          schema:
            type: string
      summary: get active promotion history by team and history name
      tags:
      - GET
  /teams/{team}/activepromotions/histories/{history}/log:
    get:
      description: Returns zip log file of the active promotion history
      parameters:
      - description: Team name
        in: path
        name: team
        required: true
        type: string
      - description: Active promotion history name
        in: path
        name: history
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: team/history should not be empty
          schema:
            type: string
        "404":
          description: activepromotion history {history} of team {team} not found
          schema:
            type: string
        "500":
          description: cannot get activepromotion history {history} of team {team}
          schema:
            type: string
      summary: Get zip log of active promotion history
      tags:
      - GET
  /teams/{team}/components:
    get:
      description: Returns list of components of team
      parameters:
      - description: Team name
        in: path
        name: team
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webhook.teamComponentsJSON'
        "404":
          description: Team not found
          schema:
            $ref: '#/definitions/webhook.errResp'
      summary: Get Team Component
      tags:
      - GET
  /teams/{team}/components/{component}/values:
    get:
      description: get team stable component values
      parameters:
      - description: Team name
        in: path
        name: team
        required: true
        type: string
      - description: Component name
        in: path
        name: component
        required: true
        type: string
      - description: Accept
        enum:
        - application/json
        - application/x-yaml
        in: header
        name: accept
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Component not found
          schema:
            $ref: '#/definitions/webhook.errResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/webhook.errResp'
      summary: get team stable component values
      tags:
      - GET
  /teams/{team}/config:
    get:
      description: get team configuration
      parameters:
      - description: Team name
        in: path
        name: team
        required: true
        type: string
      - description: Accept
        enum:
        - application/json
        - application/x-yaml
        in: header
        name: accept
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Team not found
          schema:
            $ref: '#/definitions/webhook.errResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/webhook.errResp'
      summary: get team configuration
      tags:
      - GET
  /teams/{team}/queue:
    get:
      description: Returns queue information of new component upgrading flow.
      parameters:
      - description: Team name
        in: path
        name: team
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webhook.teamQueueJSON'
        "404":
          description: Team not found
          schema:
            $ref: '#/definitions/webhook.errResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/webhook.errResp'
      summary: Get Team's Queues
      tags:
      - GET
  /teams/{team}/queue/histories/{queue}:
    get:
      description: Return queue history of team by id
      parameters:
      - description: Team name
        in: path
        name: team
        required: true
        type: string
      - description: Queue history name
        in: path
        name: queue
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1beta1.QueueHistory'
        "404":
          description: Queue history not found
          schema:
            $ref: '#/definitions/webhook.errResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/webhook.errResp'
      summary: Get Team Queue History
      tags:
      - GET
  /teams/{team}/queue/histories/{queue}/log:
    get:
      description: Returns zip log file of the queue history
      parameters:
      - description: Team name
        in: path
        name: team
        required: true
        type: string
      - description: Queue history name
        in: path
        name: queue
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1beta1.QueueHistory'
        "404":
          description: Queue history not found
          schema:
            $ref: '#/definitions/webhook.errResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/webhook.errResp'
      summary: Get Team Queue History Log
      tags:
      - GET
...
  /webhook/component:
    post:
      consumes:
      - application/json
      description: Endpoint for manually triggering new component update
      parameters:
      - description: New Component
        in: body
        name: newComponentEventJSON
        required: true
        schema:
          $ref: '#/definitions/webhook.newComponentEventJSON'
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/webhook.errResp'
      summary: Webhook New Component
      tags:
      - POST
  /webhook/github:
    post:
      description: Endpoint for receiving webhook notifications from Github
      parameters:
      - description: Github Push Event
        in: body
        name: GithubPushEvent
        required: true
        schema:
          $ref: '#/definitions/webhook.githubPushEventJSON'
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/webhook.errResp'
      summary: Github Webhook
      tags:
      - POST
swagger: "2.0"